cs50/PSet 8 Object-Oriented Programming/jar/jar.py
class Jar:
    def __init__(self, capacity=12):
        if capacity < 0:
            raise ValueError("Capacity must be at least 0")
        self._capacity = capacity
        self._size = 0


    def __str__(self):
        return "🍪" * self._size


    def deposit(self, n):
        if self._size + n > self._capacity:
            raise ValueError("Maximum capacity exceeded")
        self._size += n


    def withdraw(self, n):
        if self._size < n:
            raise ValueError("Not enough cookies to withdraw")
        self._size -= n


    @property
    def capacity(self):
        return self._capacity


    @property
    def size(self):
        return self._size